+++ src/main/java/org/jsoup/examples/Wikipedia.java
@@ -9,10 +9,13 @@ import java.io.IOException;
 
 /**
  * A simple example, used on the jsoup website.
+ <p>To invoke from the command line, assuming you've downloaded the jsoup-examples
+ jar to your current directory:</p>
+ <p><code>java -cp jsoup-examples.jar org.jsoup.examples.Wikipedia url</code></p>
  */
 public class Wikipedia {
     public static void main(String[] args) throws IOException {
-        Document doc = Jsoup.connect("http://en.wikipedia.org/").get();
+        Document doc = Jsoup.connect("https://en.wikipedia.org/").get();
         log(doc.title());
 
         Elements newsHeadlines = doc.select("#mp-itn b a");