+++ src/main/java/org/jsoup/examples/package-info.java
@@ -1,4 +1,6 @@
 /**
  Contains example programs and use of jsoup. See the <a href="https://jsoup.org/cookbook/">jsoup cookbook</a>.
+ <p>The code in this package is not available in the main `jsoup.jar` file; only in the `jsoup-examples.jar`, and has no
+ API contract. To use these examples your own project, you should copy the source code and modify as desired.</p>
  */
 package org.jsoup.examples;